+++ src/view/template.js
@@ -1,0 +1,266 @@
+import {isClassName, ucFirstName, isClassNamePopup} from '../utils.js';
+import {LENGTH_STRING, RANK} from '../mock/data.js';
+import dayjs from 'dayjs';
+import relativeTime from 'dayjs/plugin/relativeTime.js';
+
+dayjs.extend(relativeTime);
+
+export const createFilmsTemplate = () => (
+  `<section class="films">
+    <section class="films-list">
+      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+
+      <div class="films-list__container">
+
+      </div>
+    </section>
+    <section class="films-list films-list--extra">
+      <h2 class="films-list__title">Top rated</h2>
+      <div class="films-list__container">
+
+      </div>
+    </section>
+    <section class="films-list films-list--extra films-list--most-commented">
+      <h2 class="films-list__title">Most commented</h2>
+      <div class="films-list__container">
+
+      </div>
+    </section>
+  </section>`
+);
+
+const isNameRank = (count) => {
+  let nameRank = 'novice';
+  if (count >= RANK.FAN.MIN && count <= RANK.FAN.MAX) {
+    nameRank = 'fan';
+  } else if (count >= RANK.MOVIE_BUFF) {
+    nameRank = 'movie buff';
+  }
+  return ucFirstName(nameRank);
+};
+
+const isGenerateProfile = (count) => (count === 0) ? ' ' : `<p class="profile__rating">${isNameRank(count)}</p>`;
+
+export const createRankUserTemplate = (filter) => {
+  const count = filter[0].count;
+  return `<section class="header__profile profile">
+      ${isGenerateProfile(count)}
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>`;
+};
+
+
+const createMenuTemplate = (filter) => {
+  const {name, count} = filter;
+  const ucFirstNameFilter = ucFirstName(name);
+  return `<a href="#${name}" class="main-navigation__item">${ucFirstNameFilter}<span class="main-navigation__item-count">${count}</span></a>`;
+};
+
+export const createFilterTemplate = (filterItems) => {
+  const filterItemsTemplate = filterItems
+    .map((filter, index) => createMenuTemplate(filter, index === 0))
+    .join('');
+
+  return `<nav class="main-navigation">
+      <div class="main-navigation__items">
+        <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
+          ${filterItemsTemplate}
+      </div>
+      <a href="#stats" class="main-navigation__additional">Stats</a>
+    </nav>`;
+};
+
+export const createSortTemplate = () => (
+  `<ul class="sort">
+    <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
+    <li><a href="#" class="sort__button">Sort by date</a></li>
+    <li><a href="#" class="sort__button">Sort by rating</a></li>
+  </ul>`
+);
+
+
+export const createCardFilmTemplate = (card) => {
+  const {film} = card;
+  const {filmInfo, comments, userDetails} = film;
+  const {title, totalRating, releaseFilm, runtime, genre, poster, description} = filmInfo;
+  const {date} = releaseFilm;
+  const {watchList, alreadyWatched, favorite} = userDetails;
+
+  const dateYear = dayjs(date).format('YYYY');
+  const descriptionShort = (text) => text.length >= LENGTH_STRING ? `${text.substr(0, LENGTH_STRING)} <span class='more'>...</span>` : text;
+
+  return `<article class="film-card">
+    <h3 class="film-card__title">${title}</h3>
+    <p class="film-card__rating">${totalRating}</p>
+    <p class="film-card__info">
+      <span class="film-card__year">${dateYear}</span>
+      <span class="film-card__duration">${runtime}</span>
+      <span class="film-card__genre">${genre}</span>
+    </p>
+    <img src="${poster}" alt="" class="film-card__poster">
+    <p class="film-card__description">${descriptionShort(description)}</p>
+    <a class="film-card__comments">${comments.length} comments</a>
+    <div class="film-card__controls">
+      <button class="film-card__controls-item film-card__controls-item--add-to-watchlist ${isClassName(alreadyWatched)}" type="button">Add to watchlist</button>
+      <button class="film-card__controls-item film-card__controls-item--mark-as-watched ${isClassName(watchList)}" type="button">Mark as watched</button>
+      <button class="film-card__controls-item film-card__controls-item--favorite ${isClassName(favorite)}" type="button">Mark as favorite</button>
+    </div>
+  </article>`;
+};
+
+export const createShowMoreButtonTemplate = () => '<button class="films-list__show-more">Show more</button>';
+
+export const createStatisticTemplate = (length) => (
+  `<section class="footer__statistics">
+    <p>${length} movies inside</p>
+   </section>`
+);
+
+const isGenre = (array) => array.length > 1 ? 'Genres' : 'Genre';
+
+export const createPopupTemplate = (card) => {
+  const {film} = card;
+  const {filmInfo, userDetails, comments} = film;
+  const {title, totalRating, releaseFilm, genre, poster, description, ageRating, alternativeTitle, writers, director, actors, runtime} = filmInfo;
+  const {date, releaseCountry} = releaseFilm;
+  const {watchList, alreadyWatched, favorite} = userDetails;
+
+  return `<section class="film-details">
+    <form class="film-details__inner" action="" method="get">
+    <div class="film-details__top-container">
+      <div class="film-details__close">
+        <button class="film-details__close-btn" type="button">close</button>
+      </div>
+      <div class="film-details__info-wrap">
+        <div class="film-details__poster">
+          <img class="film-details__poster-img" src="${poster}" alt="">
+
+          <p class="film-details__age">${ageRating}+</p>
+        </div>
+
+        <div class="film-details__info">
+          <div class="film-details__info-head">
+            <div class="film-details__title-wrap">
+              <h3 class="film-details__title">${title}</h3>
+              <p class="film-details__title-original">Original: ${alternativeTitle}</p>
+            </div>
+
+            <div class="film-details__rating">
+              <p class="film-details__total-rating">${totalRating}</p>
+            </div>
+          </div>
+
+          <table class="film-details__table">
+            <tr class="film-details__row">
+              <td class="film-details__term">Director</td>
+              <td class="film-details__cell">${director}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Writers</td>
+              <td class="film-details__cell">${writers}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Actors</td>
+              <td class="film-details__cell">${actors}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Release Date</td>
+              <td class="film-details__cell">${date}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Runtime</td>
+              <td class="film-details__cell">${runtime}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">Country</td>
+              <td class="film-details__cell">${releaseCountry}</td>
+            </tr>
+            <tr class="film-details__row">
+              <td class="film-details__term">${isGenre(genre)}</td>
+              <td class="film-details__cell film-details__cell-genre">
+                    ${genre.join(', ')}
+              </td>
+            </tr>
+          </table>
+
+          <p class="film-details__film-description">${description}</p>
+        </div>
+      </div>
+
+      <section class="film-details__controls">
+        <button type="button" class="film-details__control-button film-details__control-button--watchlist ${isClassNamePopup(alreadyWatched)}" id="watchlist" name="watchlist">Add to watchlist</button>
+        <button type="button" class="film-details__control-button film-details__control-button--watched ${isClassNamePopup(watchList)}" id="watched " name="watched">Already watched</button>
+        <button type="button" class="film-details__control-button film-details__control-button--favorite ${isClassNamePopup(favorite)}" id="favorite" name="favorite">Add to favorites</button>
+      </section>
+    </div>
+
+    <div class="film-details__bottom-container">
+    <section class="film-details__comments-wrap">
+      <h3 class="film-details__comments-title">
+        Comments
+        <span class="film-details__comments-count">${comments.length}</span>
+      </h3>
+
+      <ul class="film-details__comments-list">
+
+      </ul>
+
+      <div class="film-details__new-comment">
+        <div class="film-details__add-emoji-label"></div>
+
+        <label class="film-details__comment-label">
+          <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
+        </label>
+
+        <div class="film-details__emoji-list">
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
+          <label class="film-details__emoji-label" for="emoji-smile">
+            <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
+          <label class="film-details__emoji-label" for="emoji-sleeping">
+            <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
+          <label class="film-details__emoji-label" for="emoji-puke">
+            <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
+          <label class="film-details__emoji-label" for="emoji-angry">
+            <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
+          </label>
+        </div>
+      </div>
+    </section>
+    </div>
+    </form>
+  </section>`;
+};
+
+export const generateCommentsList = (comments = {}) => {
+  const commentsList = document.querySelector('.film-details__comments-list');
+  for (const commentary of comments) {
+    const {author, comment, date, emotion} = commentary;
+
+    const timePassed = dayjs(date).fromNow();
+
+    const elementList = `
+    <li class="film-details__comment">
+      <span class="film-details__comment-emoji">
+        <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-smile">
+      </span>
+      <div>
+        <p class="film-details__comment-text">${comment}</p>
+        <p class="film-details__comment-info">
+          <span class="film-details__comment-author">${author}</span>
+          <span class="film-details__comment-day">${timePassed}</span>
+          <button class="film-details__comment-delete">Delete</button>
+        </p>
+      </div>
+    </li>`;
+    commentsList.insertAdjacentHTML('beforeEnd', elementList);
+  }
+};
